+++ R/utils.R
@@ -69,3 +69,33 @@ datalock <- function(x, name = NULL,
     }
   )
 }
+
+
+#' Parse and Retrieve OECD Data from URL
+#'
+#' The `fetch_oecd_data_from_url()` function is designed to facilitate easy access to OECD data by parsing a copied URL from the [OECD Data Explorer](https://data-explorer.oecd.org/). It forms a query to the database and retrieves the data in R using the `robonomistclient` package.
+#'
+#' @param url A character string representing the URL copied from the OECD Data Explorer when a particular dataset is selected.
+#' @param get A logical value indicating whether to execute the query and return the data (`TRUE`, default) or just return the query string (`FALSE`).
+#' @return If `get` is `TRUE`, the function returns the queried data. If `get` is `FALSE`, it returns the query string.
+#' @examples
+#' # Example usage:
+#' url <- ""https://data-explorer.oecd.org/vis?tm=sna&pg=0&fs[0]=Measure%2C0%7CAquaculture%20production%23AQUA_PD%23&fc=Measure&snb=1&vw=tb&df[ds]=dsDisseminateFinalDMZ&df[id]=DSD_FISH_PROD%40DF_FISH_AQUA&df[ag]=OECD.TAD.ARP&df[vs]=1.0&pd=2010%2C&dq=.A.._T.T&ly[rw]=REF_AREA&ly[cl]=TIME_PERIOD&to[TIME_PERIOD]=false""
+#' data <- fetch_oecd_data_from_url(url)
+#' query_string <- fetch_oecd_data_from_url(url, get = FALSE)
+#' @importFrom httr parse_url
+#' @importFrom glue glue
+#' @export
+fetch_oecd_data_from_url <- function(url, get = TRUE) {
+  y <- httr::parse_url(url)
+  id <- y$query$`df[id]`
+  robonomist_id <- robonomist_id("oecd", id)
+  dl_filter <- y$query$dq
+  query <- glue::glue('data_get("{robonomist_id}", dl_filter = "{dl_filter}")')
+  if (get) {
+    cat(query, "\n")
+    eval(parse(text = query))
+  } else {
+    query
+  }
+}